Extract function winEventHandler
Fix accPropServices error
